% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/raven.R
\name{provide}
\alias{provide}
\title{Provides the environment with library versions specified and runs the
code in that environment}
\usage{
provide(..., code = invisible(.libPaths()),
  local = getOption("raven.local"), repo = getOption("raven.repo"))
}
\arguments{
\item{...}{library names and versions in named list to string format
(e.g. limma='3.22.7'), if the version is ommitted (the library is
still the named parameter with no value) a warning is issued listing
available versions and the version used}

\item{code}{to run in the provisioned environment, if omitted returns
the library path of provisioned environment}

\item{local}{the local root directory for the installed packages, if
omitted uses the option 'raven.local'}

\item{repo}{the repository address, if omitted uses the option
'raven.repo'}
}
\value{
return of the code evaluated within the environment
}
\description{
Provides the environment with library versions specified and runs the
code in that environment
}

